feat: extracted formatting utils to `formatting` module
