+++ block-producer/src/server/api.rs
@@ -6,7 +6,7 @@ use miden_node_proto::{
     block_producer::api_server, requests::SubmitProvenTransactionRequest,
     responses::SubmitProvenTransactionResponse,
 };
-use miden_node_utils::logging::{format_input_notes, format_opt, format_output_notes};
+use miden_node_utils::formatting::{format_input_notes, format_opt, format_output_notes};
 use miden_objects::transaction::ProvenTransaction;
 use tonic::Status;
 use tracing::{debug, info, instrument};
