+++ block-producer/src/txqueue/mod.rs
@@ -1,7 +1,7 @@
 use std::{fmt::Debug, sync::Arc, time::Duration};
 
 use async_trait::async_trait;
-use miden_node_utils::logging::format_opt;
+use miden_node_utils::formatting::format_opt;
 use miden_objects::{
     accounts::AccountId, notes::Nullifier, transaction::InputNotes, Digest, TransactionInputError,
 };
