+++ utils/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod config;
+pub mod formatting;
 pub mod logging;
